pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "peripatos"
version = "0.1.0"
description = "Convert academic research papers into interactive Socratic audio dialogues"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "docling>=1.0.0",
    "pydub>=0.25.1",
    "edge-tts>=6.1.0",
    "openai>=1.0.0",
    "anthropic>=0.7.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

[tool.setuptools.packages.find]
include = ["peripatos*"]
exclude = ["design_guideline*", "tests*"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=6.0",
    "pytest-asyncio>=0.25",
]

[project.scripts]
peripatos = "peripatos.cli:main"

[project.urls]
Homepage = "https://github.com/Peripatos/peripatos"
Repository = "https://github.com/Peripatos/peripatos"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"